requirements.txt
pydantic
python-dotenv
dspy
attachments
ddgs
tavily-python
